refactor(ui/tiled-layout): extract tiled layout into it's own file
